+++ src/uite/__init__.py
@@ -15,7 +15,7 @@ Package metadata and version information.
 """
 
 # Package version following Semantic Versioning (SemVer)
-__version__ = "0.3.4"
+__version__ = "0.4.0"
 
 # Package authorship information
 __author__ = "Mmadubugwu Kingsley Obinna"